+++ src/test/java/demoBlazeStepDefinitions/UserRegistrationSteps.java
@@ -44,26 +44,6 @@ public class UserRegistrationSteps {
 		userRegTo.clickOkButtonInAlert();
 	}
 
-	@And("user clicks on login link")
-	public void clickOnLoginLink() {
-		userRegTo.clickOnLoginLink();
-	}
-
-	@And("user enters username {string} in login form")
-	public void enterUserNameInLoginForm(String username) {
-		userRegTo.enterUserNameInLoginForm(username);
-	}
-
-	@And("user enters password {string} in login form")
-	public void enterPasswordInLoginForm(String password) {
-		userRegTo.enterPasswordInLoginForm(password);
-	}
-
-	@Then("user clicks on login button in login form")
-	public void clickOnLoginButtonInLoginForm() {
-		userRegTo.clickOnLoginButtonInLoginForm();
-	}
-
 	@Then("user verifies error messages for sign up form")
 	public void verifyErrorMessagesForSignUpForm() {
 		userRegTo.verifyErrorMessagesForSignUpForm();
@@ -77,15 +57,5 @@ public class UserRegistrationSteps {
 	@Then("close sign up form")
 	public void closeSignUpForm() {
 		userRegTo.closeSignUpForm();
-	}
-
-	@And("user verifies wrong password error message for user login")
-	public void verifyWrongPasswordErrorMessage() {
-		userRegTo.verifyWrongPasswordErrorMessage();
-	}
-
-	@Then("close login form")
-	public void closeLoginForm() {
-		userRegTo.closeLoginForm();
 	}
 }
